rest-api/src/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-tenant-schemas==1.10.0
djangorestframework==3.12.2
ordered-set==4.0.2
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
